Move `AssertParse` tests from RecoveryTests.swift to other test files
The goal is to get rid of RecoveryTests.swift altogether once we’ve merged `XCTAssertHasSubstructure` with `AssertParse`.
